// ombi-api/model/ombi-store-entities-requests-child-requests.ts
export interface OmbiStoreEntitiesRequestsChildRequests {
    /**
     * 
     * @type {OmbiStoreEntitiesRequestsTvRequests}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    parentRequest?: OmbiStoreEntitiesRequestsTvRequests;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    parentRequestId?: number;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    issueId?: number;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    seriesType?: OmbiStoreEntitiesRequestsChildRequestsSeriesTypeEnum;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    subscribed?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    showSubscribe?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    releaseYear?: Date;
    /**
     * 
     * @type {Array<OmbiStoreEntitiesRequestsIssues>}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    issues?: Array<OmbiStoreEntitiesRequestsIssues>;
    /**
     * 
     * @type {Array<OmbiStoreRepositoryRequestsSeasonRequests>}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    seasonRequests?: Array<OmbiStoreRepositoryRequestsSeasonRequests>;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    title?: string;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    approved?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    markedAsApproved?: Date;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    requestedDate?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    available?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    markedAsAvailable?: Date;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    requestedUserId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    denied?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    markedAsDenied?: Date;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    deniedReason?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    requestType?: OmbiStoreEntitiesRequestsChildRequestsRequestTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    requestedByAlias?: string;
    /**
     * 
     * @type {OmbiStoreEntitiesOmbiUser}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    requestedUser?: OmbiStoreEntitiesOmbiUser;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    canApprove?: boolean;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesRequestsChildRequests
     */
    id?: number;
}

/**
    * @export
    * @enum {string}
    */
export enum OmbiStoreEntitiesRequestsChildRequestsSeriesTypeEnum {
    Standard = 'Standard',
    Anime = 'Anime'
}
/**
    * @export
    * @enum {string}
    */
export enum OmbiStoreEntitiesRequestsChildRequestsRequestTypeEnum {
    TvShow = 'TvShow',
    Movie = 'Movie',
    Album = 'Album'
}